Moves error handling to trait
