+++ asset-manager.php
@@ -30,6 +30,7 @@ define( 'AM_BASE_DIR', dirname( __FILE__ ) );
  * Load traits.
  */
 require_once AM_BASE_DIR . '/php/traits/trait-conditions.php';
+require_once AM_BASE_DIR . '/php/traits/trait-asset-error.php';
 
 if ( ! class_exists( 'Asset_Manager' ) ) :
 	/**
